requirements.txt
fastapi
uvicorn
Jinja2
requests
huggingface-hub
llama-cpp-python
yt-dlp
pyannote.audio
torch
torchaudio
soundfile
ffmpeg-python
openai-whisper
imageio-ffmpeg
python-dotenv
